// src/modules/auth/interfaces/IUser.ts
export interface IUser {
  id?: string;
  name?: string;
  firstName?: string;
  lastName?: string;
  username?: string;
  email?: string;
  avatar?: string;
  birthday?: string;
  instagramFollowersCount?: number;
  telegramLanguageCode?: string;
  isTelegramPremium?: boolean;
  estimatedTelegramAccountAge?: string;
  walletAddress?: string;
  ref?: string;
}